# api_docs.yaml
openapi: 3.0.0
info:
    title: Micro Service API
    version: 1.0.0
servers:
    - url: data_fetcher:5000
    - url: data_manager:5001
    - url: job_runner:5002
paths:
...
    /api/external/auth/signup:
        post:
            summary: Sign up a new user
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            type: object
                            properties:
                                email:
                                    type: string
                                password:
                                    type: string
                                name:
                                    type: string
                                PLANET_API_KEY:
                                    type: string
            responses:
                '200':
                    description: User added successfully
                '400':
                    description: User was not added
    /api/external/auth/login:
        post:
            summary: User login
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            type: object
                            properties:
                                email:
                                    type: string
                                password:
                                    type: string
            responses:
                '200':
                    description: User logged in successfully
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    user:
                                        type: object
                                        properties:
                                            _id:
                                                type: string
                                            email:
                                                type: string
                                            name:
                                                type: string
                                    token:
                                        type: string
                                    expires_in:
                                        type: integer
                '400':
                    description: Login failed
    /api/external/projects:
        get:
            summary: Get user projects
            parameters:
                - in: header
                  name: authorization
                  required: true
                  schema:
                      type: string
                - in: header
                  name: user_id
                  required: true
                  schema:
                      type: string
            responses:
                '200':
                    description: Projects retrieved successfully
                    content:
                        application/json:
                            schema:
                                type: array
                                items:
                                    type: object
                '400':
                    description: Failed to retrieve projects
        post:
            summary: Add a new project
            parameters:
                - in: header
                  name: authorization
                  required: true
                  schema:
                      type: string
                - in: header
                  name: user_id
                  required: true
                  schema:
                      type: string
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            type: object
                            properties:
                                farm_name:
                                    type: string
                                aoi:
                                    type: array
                                    items:
                                        type: array
                                        items:
                                            type: number
                                crop:
                                    type: string
                                created_at:
                                    type: string
                                seeding_date:
                                    type: string
            responses:
                '200':
                    description: Project added successfully
                '400':
                    description: Failed to add project
...
    /api/external/projects/{project_id}/image:
        post:
            summary: Get image at date for a project
            parameters:
                - in: path
                  name: project_id
                  required: true
                  schema:
                      type: string
                - in: header
                  name: authorization
                  required: true
                  schema:
                      type: string
                - in: header
                  name: user_id
                  required: true
                  schema:
                      type: string
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            type: object
                            properties:
                                date_of_interest:
                                    type: string
            responses:
                '200':
                    description: Image retrieved successfully
                    content:
                        image/png:
                            schema:
                                type: string
                                format: binary
                '400':
                    description: Bad request
                '401':
                    description: Unauthorized
                '404':
                    description: Project not found
...